fix coordinates on edit modal for new metaverse
