--- src/modules/land-works/components/land-works-selected-feature-card-v2/s.module.scss
@@ -1,39 +1,0 @@
-
-@import '../../../../styles/utils.scss';
-
-.wrapper {
-  max-width: 406px;
-  width: 100%;
-  @media all and (max-width: 760px) {
-    margin-bottom: 20px;
-  }
-
-  .card {
-    cursor: pointer;
-    border-radius: 10px;
-    border: 2px solid transparent;
-    transition: all 0.2s ease-out;
-    padding: 10px;
-    box-shadow: none;
-    background: var(--theme-modal-color);
-
-    .details {
-      color: var(--theme-grey700-color);
-      font-weight: 600;
-      font-size: 14px;
-    }
-
-    .card:hover {
-      box-shadow: 0px 5px 18px rgba(255, 255, 255, 0.3);
-      border: 2px solid var(--theme-primary-color);
-    }
-
-    .name {
-      font-size: 16px;
-      font-weight: bold;
-      color: var(--theme-primary-color);
-      text-transform: uppercase;
-      margin: 8px 0 5px;
-    }
-  }
-}
